move useeffect
