+++ webapp/components/form/AutocompleteDialog/AutocompleteDialog.js
@@ -1,6 +1,6 @@
 import './AutocompleteDialog.scss'
 
-import React, { useRef, useEffect, useMemo } from 'react'
+import React, { useRef, useMemo } from 'react'
 import PropTypes from 'prop-types'
 
 import { useAutocompleteDialog, State } from './store'
@@ -26,37 +26,13 @@ const AutocompleteDialog = (props) => {
     itemKey,
     onItemSelect,
     onClose,
+    list,
   })
 
   const calculatedPosition = useMemo(() => State.calculatePosition(state), [
     State.getSourceElement(state),
     State.getInputField(state),
   ])
-
-  useEffect(() => {
-    const keyDownListener = Actions.onInputFieldKeyDown({
-      list,
-      state,
-    })
-
-    const clickListener = Actions.onOutsideClick({
-      list,
-      state,
-    })
-
-    if (inputField) {
-      inputField.addEventListener('keydown', keyDownListener)
-    }
-    window.addEventListener('click', clickListener)
-
-    return () => {
-      if (inputField) {
-        inputField.removeEventListener('keydown', keyDownListener)
-      }
-
-      window.removeEventListener('click', clickListener)
-    }
-  }, [State.getInputField(state), list.current])
 
   return (
     <div ref={list} className={`autocomplete-list ${className}`} style={{ ...calculatedPosition }}>
