+++ webapp/components/form/AutocompleteDialog/store/useAutocompleteDialog.js
@@ -1,4 +1,4 @@
-import { useState } from 'react'
+import { useEffect, useState } from 'react'
 
 import { useActions } from './actions'
 import { State } from './state'
@@ -11,6 +11,7 @@ export const useAutocompleteDialog = ({
   itemKey,
   onItemSelect,
   onClose,
+  list,
 }) => {
   const [state, setState] = useState(() =>
     State.create({
@@ -22,6 +23,31 @@ export const useAutocompleteDialog = ({
     })
   )
   const Actions = useActions({ setState, onItemSelect, onClose })
+
+  useEffect(() => {
+    const keyDownListener = Actions.onInputFieldKeyDown({
+      list,
+      state,
+    })
+
+    const clickListener = Actions.onOutsideClick({
+      list,
+      state,
+    })
+
+    if (inputField) {
+      inputField.addEventListener('keydown', keyDownListener)
+    }
+    window.addEventListener('click', clickListener)
+
+    return () => {
+      if (inputField) {
+        inputField.removeEventListener('keydown', keyDownListener)
+      }
+
+      window.removeEventListener('click', clickListener)
+    }
+  }, [State.getInputField(state), list.current])
 
   return { Actions, state }
 }
